style: apply `just audit --autofix` across the integration tree
Mechanical cleanup from the sanctioned autofix pipeline: clippy --fix
idioms, ast-grep --fix, xtask lint --fix (impl-item sort + comment
hygiene), fmt. Net reduction in arch-ratchet violations (file_size,
god_trait, pub_struct_open_fields, single_impl_size, single_word_filenames
all 1 -> 0). No logic changes; lint-fast green (53 deny, 561 warn, 0 new,
0 regression).

+++ tests/tests/phase_continuity/file.rs
@@ -531,7 +531,7 @@ async fn dump_aac_for_listening() {
             / residual.len() as f64)
             .sqrt();
         let snr_db = if rms > 1e-9 {
-            20.0 * (f64::from(amp) / rms).log10()
+            20.0 * (amp / rms).log10()
         } else {
             f64::INFINITY
         };
